don't move blocks that are locked in.
This seems to be the better way of finding locked in blocks. This
will prevent awkward holes and change slopes to look a bit more
usual during erosion.

To test this I wrote a test function that allows me to quickly
do a lot of erosion attempts on a small section of the map, which
will be handy since I've been having a hard time testing the
validity of the algorithms due to the math scale of things (even
blocks of sand were resisting 1000+ attempts to move down 1 spot!).
Having a hammer to whack at it hard will reduce my testing time
and will make for an easier time tuning the parameters.