Initial work for glaciers.
This code will create growing and thawing snow. The snow will grow
if it's storming and temperatures are low, and snow will melt if
the sun is out or some other light source is near. The algorithm
makes very pleasing rounded mountain tops.

From here on, we need to pile snow higher to produce ice blocks, the
next step needed to create glacier-like effects.