Attempt to make waterfalls behave more natural
With the old code, waterfalls would only ever grow and
the amount of water blocks would never decrease. This
leads to huge pillars of water that are unrealistic and
create enormously large water fans, which is ugly at
best.

This code removes the water column entirely above blocks
that are moved. By doing so, we cause a waterfall block
update which will likely put the water back again, but
most of the time do so one block lower. This causes the
water depth to be adjusted continuously lower, and leads
to a more realistic flow.

Ultimately, this code will also remove the waterfall
itself since it will remove the water origin block as
well if it is no longer protected. Since the landscape
generator usually put waterfalls *under* another stone
block, this will take a long time. All water will disappear
and you're left with a "dry" valley and alluvial fan below,
which bears resemblance to dry valleys in nature.