Snow accumulation rewrite.
No longer recursive, and tweaked all the parameters to be
a bit simpler and easier to check. The end result is a lot
better than before, since I broke quite a few aspects as
I was going through this code.

Pretty smooth snowy hilltops, growth is smooth as well.

Thawing is done right at the start which simplifies and
makes things a bit faster.